Make conf JSON handling more robust.
Do some simple checks on a list of required fields and a dictionary of
optional fields (and their associated default values) to handle
variations in the JSON schema supplied by the user gracefully.

For the most part, fixes Issue #6, unless the configuration file becomes
very complicated in the near future. At that point, adding a dependency
for jsonschema is probably the move. For now, we won't add an entire
module to support functionality that can be done in a handful of lines.