Make state.py CmdError more readable to end-user.
The original error message was pretty cryptic, especially if you don't
know the code. However, this error was being shown to end-users. A
better error message is now shown.

This commit finalizes the resolution of Issue #3.

The only complaint I have is that I am not giving an error message
specifically about the config file's project name potentially being
incorrect. It just defaults to the standard invalid argument error. This
should be fine for now.