Add better error message printing and comment handling.
This also improves on Issue #3's resolution by allowing us to inject a
comment into the project selection command that marks it as a command
not executed by the user but by the program in an attempt to honor the
user's default_project request.

The error message printing has been improved to print out the error
itself, the command that caused it, and its arguments that were given.

Comment handling only respects #-marked comments.