Change error to refer to project validity, not id
The error for an incorrect select command would suggest that the
argument gave an invalid project id. While this is technically correct,
it is possible that the user opted to use something like string
substitution (select %s:"Inbox"). If this is done, then it doesn't make
immediate sense why the project id was wrong, unless you know how the
back-end is substituting "Inbox" for the actual project id. Using 'valid
project' is much more general and implies that the actual string
substitution has something wrong with it, like a typo.