Add default case for pat, hint creation
We were missing an else clause in _get_pat_and_hint(arg). This meant
that if we were on a command token that had no %<character>:<value>,
then we would throw exceptions. An example of this being possible is for
task complete %p:#. In this case, when we were looking to inject on the
token 'complete', we would throw an error since we would try to return a
pat and hint before we had assigned it any values (since none of the
cases would execute). The else clause now sets pat and hint to None as
they would have been in the older code.