Properly return values from decorators
We were not returning values properly from decorators, causing some of
the methods on TodoistCLI() to actually return None. This was okay just
because their return values seem to never really be used so this bug was
not discovered. I would not be surprised if this kind of bug exists
elsewhere for other methods of TodoistCLI(), but for now, I've fixed
this one.