Fix issue where returning from do_cmd() halts cmdloop
This is actually documented in the Cmd module's documentation, but I did
not read this when implementing this initially. This problem was then
masked because my decorators were not "properly" returning the return
values of the functions they were wrapping. Once I "fixed" that in an
earlier commit, it surfaced this new problem, which we fix by overriding
onecmd() to not act so stupidly (I still have _no_ idea why Cmd thinks
it makes sense to terminate its entire loop just because a do_cmd() has
returned something truthy instead of doing something more reasonable
like halting when a do_cmd() raises an error).